@@ -12,13 +12,13 @@
     "minimum-stability": "stable",
     "require": {
         "php": ">=7.0",
-        "laravel/framework": "5.4.*",
+        "laravel/framework": "5.5.*",
         "maatwebsite/excel": "~2.1.0"
     },
     "require-dev":{
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~5.7"
+        "phpunit/phpunit": "~6.0"
     },
     "autoload": {
         "psr-4": {
